flask_sqlmodel_Registration with HTMX/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9
python-dotenv==1.1.0
